nmt_model/requirements.txt
absl-py==0.8.0
attrs==19.2.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
ConfigArgParse==0.14.0
conllu==2.2
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
Flask==1.1.1
future==0.17.1
grpcio==1.24.0
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonpickle==0.9.6
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
mpmath==1.1.0
munch==2.3.2
nbconvert==5.6.0
nbformat==4.4.0
nose==1.3.7
notebook==6.0.1
numpy==1.17.2
OpenNMT-py==1.0.0rc2
packaging==19.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.9.2
ptyprocess==0.6.0
py-cpuinfo==5.0.0
Pygments==2.4.2
pymongo==3.9.0
pyonmttok==1.15.7
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-Levenshtein==0.12.0
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
rouge==0.3.2
sacred==0.7.5
Send2Trash==1.5.0
sentencepiece==0.1.83
six==1.12.0
soupsieve==1.9.4
sympy==1.4
tensorboard==2.0.0
terminado==0.8.2
testpath==0.4.2
torch==1.2.0
torchtext==0.4.0
tornado==6.0.3
tqdm==4.30.0
traitlets==4.3.3
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
